mcux: sdk-ng: update device/i.mx to 26.09.00 RC1
Sync the i.MX device headers, feature data, and generated support files
with the MCUX SDK 26.09.00 RC1 content. The sync keeps the hal_nxp
cleanup exclusions for unused large TRDC and NOC GIC headers.

+++ mcux/mcux-sdk-ng/devices/i.MX/i.MX95/MIMX9596/drivers/fsl_iomuxc.h
@@ -948,12 +948,12 @@ static inline void IOMUXC_SetPinMux(uint32_t muxRegister,
 {
     if (muxRegister != 0U)
     {
-        *((volatile uint32_t *)muxRegister) = IOMUXC_PAD_MUX_MODE(muxMode) | IOMUXC_PAD_SION(inputOnfield);
+        *((volatile uint32_t *)(uintptr_t)muxRegister) = IOMUXC_PAD_MUX_MODE(muxMode) | IOMUXC_PAD_SION(inputOnfield);
     }
 
     if (inputRegister != 0U)
     {
-        *((volatile uint32_t *)inputRegister) = inputDaisy;
+        *((volatile uint32_t *)(uintptr_t)inputRegister) = inputDaisy;
     }
 }
 /*!
@@ -976,7 +976,7 @@ static inline void IOMUXC_SetPinConfig(uint32_t muxRegister,
 {
     if (configRegister != 0U)
     {
-        *((volatile uint32_t *)configRegister) = configValue;
+        *((volatile uint32_t *)(uintptr_t)configRegister) = configValue;
     }
 }
 /*@}*/
